Use npm instead of yarn for global package installation
As yarn maintains a flat dependency structure, each time a global
package is added / removed it will examine all other global packages
to adjust dependencies accordingly. It is desired for per-project
modules but not necessary for global packages. Since it is typical
to add / remove global packages one by one, the time spent on the
dep analysis will defeat the purpose to use yarn.